+++ NEWS.md
@@ -1,3 +1,11 @@
+# bs4Dash 0.3.0.9000
+## Bug fix
+- fix #8: Compatibility with shinyWidgets pickerInput. Thanks @pvictor.
+- fix #13 in `bs4TabItem`: handle tabName with space. Thanks @stephLH.
+- fix `bs4TabCard`: when the user set 2 or more tabs active at the same time, only 
+the first one is selected
+
+
 # bs4Dash 0.2.0
 ## Major changes
 - bs4Dash now works on shinyapp.io (previously, there was a problem with tab navigation)